Do a better job sorting stable releases
In preparation for a potential 4.0 (just in case Linus decides to go
ahead with that). Existing code only sorted by the second version
number, so 3.19.1 was sorted above 4.0.1. With this change, we use a
slower but a more sure mechanism of pretending these are floats and
comparing .3191 and .401, which should always sort in correct order.