Rewrite signatures doc to use TOFU and WKD
People who are likely to be reading the Signatures doc are not going to
be part of the kernel.org web of trust, so telling them to use the web
of trust to verify release signatures is disingenious.

Recommend using TOFU and the WKD for obtaining public keys, and document
the procedure (requires gnupg-2.2).

Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>